+++ storybook/stories/Furniture.stories.ts
@@ -49,6 +49,7 @@ export function Default() {
       roomZ: 0,
       type: "edice",
       direction: 0,
+      animation: "0",
     });
 
     dice.extradata.then(action("Extra Data"));
@@ -323,47 +324,95 @@ export function DifferentFetchTypes() {
 export function DetachedFurniture() {
   return createShroom(({ application, shroom }) => {
     const container = new PIXI.Container();
+    container.sortableChildren = true;
+
     application.stage.addChild(container);
 
     const furniture = BaseFurniture.fromShroom(shroom, container, {
       animation: "0",
-      direction: 0,
+      direction: 2,
       type: { type: "club_sofa", kind: "type" },
     });
+
+    furniture.x = 100;
+    furniture.y = 100;
   });
 }
 
-export function PlacementFloorFurnitureTest() {
+export function GldGate() {
   return createShroom(({ application, shroom }) => {
     const container = new PIXI.Container();
     application.stage.addChild(container);
 
     const room = Room.create(shroom, {
       tilemap: `
-       xxxxxxxxxxx
-       x0000000000
-       x0000000000
-       x0000000000
-       x0000000000
-       x0000000000
-       x0000000000
-       x0000000000
-       x0000000000
-      `,
-    });
-
-    const furniture = new FloorFurniture({
-      roomX: 0,
-      roomY: 0,
-      roomZ: 0,
-      animation: "0",
-      direction: 2,
-      type: "club_sofa",
+           xxxxxxxxxxx
+           x0000000000
+           x0000000000
+           x0000000000
+           x0000000000
+           x0000000000
+           x0000000000
+           x0000000000
+           x0000000000
+          `,
     });
 
     room.x = application.screen.width / 2 - room.roomWidth / 2;
     room.y = application.screen.height / 2 - room.roomHeight / 2;
-    room.addRoomObject(furniture);
+
+    const floorFurniture = new FloorFurniture({
+      type: "gld_gate",
+      direction: 2,
+      roomX: 1,
+      roomY: 1,
+      roomZ: 0,
+      animation: "0",
+    });
+
+    const floorFurniture2 = new FloorFurniture({
+      type: "gld_gate",
+      direction: 2,
+      roomX: 1,
+      roomY: 3,
+      roomZ: 0,
+      animation: "1",
+    });
+
+    let b = false;
+
+    const floorFurniture3 = new FloorFurniture({
+      type: "gld_gate",
+      direction: 2,
+      roomX: 1,
+      roomY: 5,
+      roomZ: 0,
+      animation: "0",
+    });
+
+    floorFurniture3.onClick = () => {
+      b = !b;
+
+      if (b) {
+        floorFurniture3.animation = "100";
+      } else {
+        floorFurniture3.animation = "101";
+      }
+    };
+
+    const floorFurniture4 = new FloorFurniture({
+      type: "gld_gate",
+      direction: 2,
+      roomX: 1,
+      roomY: 7,
+      roomZ: 0,
+      animation: "101",
+    });
+
+    //room.addRoomObject(floorFurniture);
+    //room.addRoomObject(floorFurniture2);
+    room.addRoomObject(floorFurniture3);
+    //room.addRoomObject(floorFurniture4);
 
     application.stage.addChild(room);
   });